all tests are relocated into a new root folder __tests__ to correspond to the Jest convention
